pyproject.toml
[tool.poetry]
name = "iot-dashboard"
version = "0.1.0"
description = ""
authors = ["Vladislav <Vladislavzikstudy@gmail.com>"]
readme = "README.md"
packages = [{include = "iot_dashboard"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
djoser = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
